+++ nginx_ldap_auth/exc.py
@@ -1,0 +1,4 @@
+class ImproperlyConfigured(Exception):
+    """
+    Raised when the application is improperly configured.
+    """